% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/booleanize_clusters_function.R
\name{booleanize_clusters}
\alias{booleanize_clusters}
\title{Function to booleanize gene expression values in subpopulations based on clustering}
\usage{
booleanize_clusters(fluidSCproc, based_on_values = "log2ExNorm",
  clusterColumn, boolean_method = c("Wilcox_test", "Kmeans_2_clusters"),
  Wilcox_pval = 0.4, print_heatmap = T)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{clusterColumn}{name of column that contains cluster data}

\item{boolean_method}{create boolean values based on Wilcoxon test or Kmeans clustering of each assay}

\item{Wilcox_pval}{p-value treshold to classify genes as active or inactive based on one-sided Wilcoxon test ("greater")}

\item{print_heatmap}{print heatmap from boolean dataframe endresult}
}
\value{
returns a dataframe with boolean values for all subpopulations
}
\description{
Create boolean values in subpopulations based on non-parametric Wilcoxon test or with Kmeans clustering (active and inactive cluster)
}
\details{
NA
}
\examples{
booleanize_clusters()
}

