% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_sparse_SC_function.R
\name{cluster_sparse_SC}
\alias{cluster_sparse_SC}
\title{Function to sparse cluster single cell data using the sparcl package}
\usage{
cluster_sparse_SC(fluidSCproc, based_on_values = "log2ExNorm",
  scaleData = F, sparseClust = c("Hierarchical", "Kmeans"), wbound = 4,
  cluster_column_name = "clust", nrClust = 2, return_geneWeights = F,
  NAvalues = c("remove_assays", "replace_with_not_detected"),
  not_detected_value = 0, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{scaleData}{boolean to scale data before clustering, default to F}

\item{sparseClust}{choose method for clustering from sparcl package}

\item{wbound}{parameter for sparse clustering that determines feature weights}

\item{cluster_column_name}{name of column that will be created with the clustering results}

\item{nrClust}{number of clusters you want to detect (K for sparse kmeans clustering)}

\item{return_geneWeights}{return weights of genes to clustering instead of fluidSCproc object}

\item{NAvalues}{how to handle NA values, remove or replace with not_detected_value}

\item{not_detected_value}{value to replace NA values}

\item{...}{additional parameters for HierarchicalSparseCluster or KmeansSparseCluster from sparcle package}
}
\value{
returns a fluidSCproc S3 object appended with the cluster results OR a vector with gene weights if return_geneWeights = TRUE
}
\description{
This function is a wrapper for the sparcl package functions HierarchicalSparseCluster and KMeansSparseCluster and appends the result to the fluidSCproc object
or gives a vector with gene weights for clustering of the data
}
\details{
NA
}
\examples{
cluster_sparse_SC()
}

