% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_sparse_SC_function.R
\name{cluster_sparse_SC_permute}
\alias{cluster_sparse_SC_permute}
\title{Function to determine optimal wbound parameter for cluster_sparse_SC function}
\usage{
cluster_sparse_SC_permute(fluidSCproc, based_on_values = "log2ExNorm",
  scaleData = F, sparseClust = c("Hierarchical", "Kmeans"), nperms = 10,
  wbounds = c(1.5, 2:6), K = NULL, NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{scaleData}{boolean to scale data before clustering, default to F}

\item{sparseClust}{choose method for clustering from sparcl package}

\item{nperms}{number of permutations}

\item{wbounds}{vector of values for sparse clustering that determines feature weights}

\item{K}{number of cluster if chosen Kmeans as clustering method}

\item{NAvalues}{how to handle NA values, remove or replace with not_detected_value}

\item{not_detected_value}{value to replace NA values}

\item{...}{additional parameters for HierarchicalSparseCluster.permute or KmeansSparseCluster.permute from sparcle package}
}
\value{
returns info about the optimal wbound parameter
}
\description{
This function is a wrapper for the sparcl package functions HierarchicalSparseCluster.permute and KMeansSparseCluster.permute
It will given an indication about the wbound parameter
}
\details{
NA
}
\examples{
cluster_sparse_SC_permute()
}

