% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display_parameters_functions.R
\name{display_parameters_distribution}
\alias{display_parameters_distribution}
\title{Function to output parameters of distribution}
\usage{
display_parameters_distribution(fluidSCraw, reference_genes = NULL,
  groupColumn = NULL)
}
\arguments{
\item{reference_genes}{reference genes in your assays}

\item{groupColumn}{create facets based on column with grouping}

\item{rawFluidCt}{data slot from S3 fluidSCraw or fluidSCproc object}
}
\value{
outputs text in console with information
}
\description{
This function outputs distribution parameters like range, quantile, median, ... calls display_parameters_function()
}
\examples{
display_parameters_distribution()
}

