% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/display_parameters_functions.R
\name{display_parameters_function}
\alias{display_parameters_function}
\title{Function to calculate parameters of distribution}
\usage{
display_parameters_function(rawFluidCt, reference_genes = NULL,
  groupColumn = NULL)
}
\arguments{
\item{rawFluidCt}{data slot from S3 fluidSCraw or fluidSCproc object}

\item{reference_genes}{reference genes in your assays}

\item{groupColumn}{create facets based on column with grouping}
}
\value{
outputs text in console with information
}
\description{
This function calculates distribution parameters like range, quantile, median, ...
}
\examples{
display_parameters_function()
}

