% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dotplotSC_function.R
\name{dotplotSC}
\alias{dotplotSC}
\title{dotplotSC}
\usage{
dotplotSC(fluidSCproc, based_on_values = "log2Ex", assayCol = "Assays",
  assays = c("Rpl13a", "Pou5f1"), missingCol = NULL, coord_ratio = 0.6,
  binwidth = 0.6)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{assayCol}{column name with the assays, defaults to "Assays"}

\item{assays}{vector with assays to show}

\item{missingCol}{optional column name with information for missing values (or other features)}

\item{coord_ratio/binwidth}{graphical parameters that need to be adjusted for proper visualization (should have the same value ideally)}
}
\value{
returns a dotplot
}
\description{
Function to visualize distribution profile of assays in single cells with a dotplot
}
\details{
NA
}
\examples{
dotplotSC()
}

