% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expression_cluster_SC_function.R
\name{expression_cluster_SC}
\alias{expression_cluster_SC}
\title{Function to to create barplots from different cluster groups}
\usage{
expression_cluster_SC(fluidSCproc, based_on_values = "log2ExNorm",
  cluster_column_name = "clust", centrality_function = mean,
  log_base_adjust = 10, cluster_genes = c("Hierarchical", "Correlation"),
  nrClust = 6, distMethod = "euclidean", clustMethod = "average",
  geneWeights = NULL, quantile_breaks = c(0.25, 0.5, 0.75, 0.95, 1),
  quantile_labels = c("very low", "low", "medium", "high", "very high"),
  alpha_range = c(0.5, 1), selected_assays = NULL)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{cluster_column_name}{name of the column with cluster information}

\item{centrality_function}{function to use for calculating a centrality score, e.g. mean, median, ...}

\item{log_base_adjust}{transform deviation scores}

\item{cluster_genes}{algorithm to cluster assays}

\item{nrClust}{number of assay clusters}

\item{distMethod}{method to calculate distance matrix}

\item{clustMethod}{method to cluster distance matrix}

\item{geneWeights}{named assay vector with weights, will be used to highlight more important genes}

\item{quantile_breaks}{divide geneWeights in quantile groups}

\item{quantile_labels}{give labels to groups created by quantile_breaks}

\item{alpha_range}{vector with range for alpha-value}

\item{selected_assays}{select subset of assays}
}
\value{
returns a barplot with facets for different clusters and important assays highlighted based on assay weights
}
\description{
This function will calculate deviation scores for the different cluster groups from the population average and display them as bar plots.
In addition it will also cluster the assays.
}
\details{
NA
}
\examples{
expression_cluster_SC()
}

