% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructor_S3_fluidSCproc.R
\name{fluidSCproc}
\alias{fluidSCproc}
\title{Constructor function for fluidSCproc S3 object from dataframe}
\usage{
fluidSCproc(x, LoD = NULL)
}
\arguments{
\item{x}{dataframe with at least 5 columns: Samples, Assays, Value, Call and log2Ex}

\item{LoD}{need to provide Limit of Detection (LoD) that you used to calculate log2 expression values (log2Ex)}
}
\value{
S3 object fluidSCproc for analysis and visualization
}
\description{
This function creates a fluidSCproc S3 object.
}
\examples{
fluidSCproc()
}
\keyword{construct}
\keyword{fluidSCproc}

