% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fluidSCproc_to_newCellDataSet_function.R
\name{fluidSCproc_to_newCellDataSet}
\alias{fluidSCproc_to_newCellDataSet}
\title{Function to create a newCellDataSet object from fluidSCproc object to use in the monocle package}
\usage{
fluidSCproc_to_newCellDataSet(fluidSCproc,
  based_on_values = "log2ExNormMerge", sampleData = NULL,
  featureData = NULL, NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2ExNorm"}

\item{sampleData}{character vector with column names that contain information about the samples}

\item{featureData}{character vector with column names that contain information about the features (assays)}

\item{NAvalues}{how to process NA values, remove or replace with not_detected_value (default = 0)}

\item{not_detected_value}{value to replace NA values with}
}
\value{
returns a t-SNE plot
}
\description{
This function creates newCellDatqSet object. This is the starting object for the monocle package.
}
\details{
NA
}
\examples{
fluidSCproc_to_newCellDataSet()
}
\references{
http://monocle-bio.sourceforge.net/tutorial.html
}

