% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructor_S3_fluidSCraw.R
\name{fluidSCraw}
\alias{fluidSCraw}
\title{Constructor function for fluidSCraw S3 object from dataframe}
\usage{
fluidSCraw(x)
}
\arguments{
\item{x}{dataframe with at least 4 columns: Samples, Assays, Value and Call}
}
\value{
S3 object fluidSCraw for pre-processing steps
}
\description{
This function creates a fluidSCraw S3 object.
}
\examples{
fluidSCraw()
}
\keyword{construct}
\keyword{fluidSCraw}

