% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifyBadAssays_function.R
\name{identifyBadAssays}
\alias{identifyBadAssays}
\title{Function to identifie bad assays}
\usage{
identifyBadAssays(fluidSCraw, min_ratio_goodVSbad = 1)
}
\arguments{
\item{fluidSCraw}{fluidSCraw S3 object}

\item{min_ratio_goodVSbad}{ratio of number of good calls over bad calls, ratio = #good/#bad. Defaults to 1, i.e. minimum half of the calls must be good}
}
\value{
returns list with overview of valued and result, i.e. names of assays that do not meet the requirements
}
\description{
This function determines if assays are of good quality based on the ratio of good versus bad quality calls
}
\examples{
identifyBadAssays()
}

