% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifyBadSamples_function.R
\name{identifyBadSamples}
\alias{identifyBadSamples}
\title{Function to identify bad samples}
\usage{
identifyBadSamples(fluidSCraw, sizefactor = 3)
}
\arguments{
\item{fluidSCraw}{fluidSCraw S3 object}

\item{sizefactor}{factor to multiply median missing data value, defaults to 3 see details}
}
\value{
returns character vector with names of bad Samples
}
\description{
This function determines the quality of the samples based on an indication of missing data.
The more an assay is expressed in all samples (expression probability), the heavier its weight.
}
\details{
sizefactor 1st step) calculate expression probability per assay
2nd step) failed calls get the value of this expression probability and are summed per sample
3th step) if sum of sample i > median for all samples * sizefactor => consider it a bad sample
}
\examples{
identifyBadAssays()
}
\references{
Livak et al. "Methods for qPCR gene expression profiling applied to 1440 lymphoblastoid single cells"
}

