% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/identifyOutlierSamples_function.R
\name{identifyOutlierSamples}
\alias{identifyOutlierSamples}
\title{Function to identify possible outlier samples}
\usage{
identifyOutlierSamples(fluidSCproc, based_on_values = "log2Ex",
  color_outliers_probability = 0.05)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use for IAC score, defaults to "log2Ex"}

\item{color_outliers_probability}{cutoff to color outliers based on quantile probability distribution}
}
\value{
returns a boxplot with sample names and possible outliers colored in red
}
\description{
This function searches for possible outliers based on an Inter Array Correlation (IAC) score and highlight these based on quantile probability score
}
\details{
NA
}
\examples{
identifyOutlierSamples()
}
\references{
Michael C. Oldham et al. "Functional Organization of the Transcriptome in Human Brain" Supplement
}

