% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kmeans_cluster_estimation_functions.R
\name{Kmeans_cluster_explained_variance}
\alias{Kmeans_cluster_explained_variance}
\title{Function to visualize explained variance by different Kmeans cluster. Creates scree (elbow) plot.}
\usage{
Kmeans_cluster_explained_variance(fluidSCproc, based_on_values = "log2Ex",
  test_nrClust = 10, selected_assays = NULL, NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{test_nrClust}{max number of clusters to test}

\item{selected_assays}{make selection of assays if wanted}

\item{NAvalues}{how to handle NA values, remove or replace with not_detected_value}

\item{not_detected_value}{value to replace NA values}

\item{...}{additional parameters for HierarchicalSparseCluster.permute or KmeansSparseCluster.permute from sparcle package}
}
\value{
returns info about the optimal wbound parameter
}
\description{
This function will perform Kmeans clustering on your fluidSCproc object for a range of defined clusters (1 to test_nrClust) and
the output will be a scree-plot, where you see the additional variance that is explained with each cluster.
}
\details{
NA
}
\examples{
Kmeans_cluster_explained_variance()
}

