% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kmeans_cluster_estimation_functions.R
\name{Kmeans_cluster_statistics}
\alias{Kmeans_cluster_statistics}
\title{Function to determine best number of clusters in kmeans clustering}
\usage{
Kmeans_cluster_statistics(fluidSCproc, based_on_values = "log2Ex",
  scaleData = F, test_nrClust = 10, method = c("Kmeans", "sparseKmeans"),
  wbound = 2, selected_assays = NULL, NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0, clusGap_B = 100,
  clusterboot_B = 100, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{test_nrClust}{max number of clusters to test}

\item{selected_assays}{make selection of assays if wanted}

\item{NAvalues}{how to handle NA values, remove or replace with not_detected_value}

\item{not_detected_value}{value to replace NA values}

\item{clusGap_B}{number of monte carlo ("bootstrap") samples for gap statistic}

\item{clusterboot_B}{Number of resampling runs for each scheme for bootstrap method, see clusterboot}
}
\value{
returns and prints estimations for optimal number of clusters based on 2 different "bootstrap" algorithms (functions)
}
\description{
This function will calculate gap-statistics for range of K clusters and it will also output results of a bootstrap function on range of kmeans clustering results.
}
\details{
NA
}
\examples{
Kmeans_cluster_statistics()
}

