% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanExpression_vs_standardDeviation_SC_function.R
\name{meanExpression_vs_standardDeviation_SC}
\alias{meanExpression_vs_standardDeviation_SC}
\title{meanExpression_vs_standardDeviation_SC}
\usage{
meanExpression_vs_standardDeviation_SC(fluidSCproc,
  based_on_values = "lin2Ex", selected_genes = NULL, log_base = NULL,
  label_deviation = 0.7, show_labels = T, log2Ex_maximum_COV = T,
  log2Ex_COV_abline = 0.25)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "lin2Ex"}

\item{selected_genes}{subset of genes}

\item{log_base}{transform values with log}

\item{label_deviation}{jitter on the y-axis for the labels}

\item{show_labels}{boolean to show labels, defaults to TRUE}

\item{log2Ex_maximum_COV}{boolean to draw line with hypothetical maxium Coefficient of Variation (COV)}

\item{log2Ex_COV_abline}{value to draw a line with constant COV on plot, defaults to 0.25 (25 %)}
}
\value{
returns a list with plot and corresponding data
}
\description{
Function to get and plot mean abundance versus standard deviation of expression in single cells
}
\details{
NA
}
\examples{
meanExpression_vs_standardDeviation_SC()
}

