% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanExpression_vs_variance_SC_function.R
\name{meanExpression_vs_variance_SC}
\alias{meanExpression_vs_variance_SC}
\title{meanExpression_vs_variance_SC}
\usage{
meanExpression_vs_variance_SC(fluidSCproc, based_on_values = "lin2Ex",
  log_base = NULL, label_deviation = 0.7, show_labels = T)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "lin2Ex"}

\item{log_base}{transform values with log}

\item{label_deviation}{jitter on the y-axis for the labels}

\item{show_labels}{boolean to show labels, defaults to TRUE}
}
\value{
returns a list with plot and corresponding data
}
\description{
Function to get and plot mean abundance versus variance of expression in single cells
}
\details{
NA
}
\examples{
meanExpression_vs_variance_SC()
}

