% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_fluidSCraw_function.R
\name{merge_fluidSCraw}
\alias{merge_fluidSCraw}
\title{Function to merge fluidSCraw objects}
\usage{
merge_fluidSCraw(fluidSCraw_list)
}
\arguments{
\item{fluidSCraw_list}{list of fluidSCraw S3 objects, need to have the same column names}
}
\value{
returns a fluidSCraw S3 object
}
\description{
This function merges multiple fluidSCraw S3 objects
}
\examples{
merge_fluidSCraw()
}

