% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_normalized_fluidSC_function.R
\name{merge_normalized_fluidSC}
\alias{merge_normalized_fluidSC}
\title{Function to merge normalized fluidSCproc objects}
\usage{
merge_normalized_fluidSC(fluidSCproc_list)
}
\arguments{
\item{fluidSCproc_list}{list of fluidSCproc S3 objects, need to have the same column names}
}
\value{
returns a fluidSCproc S3 object
}
\description{
This function merges multiple fluidSCproc S3 objects and creates a new cell-to-cell median normalization log2 expression value for combined samples
}
\details{
Normalization is based on cell-to-cell median. Offset value between median log2 expression in sample vs average of all median values.
End result is that all cells will have the same median log2 expression.
}
\examples{
merge_normalized_fluidSC()
}
\references{
Livak et al. "Methods for qPCR gene expression profiling applied to 1440 lymphoblastoid single cells"
}

