% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multi_IQRexpression_range_function.R
\name{multi_IQRexpression_range}
\alias{multi_IQRexpression_range}
\title{Function to show expression range of genes based on quantile probabilities for multiple conditions}
\usage{
multi_IQRexpression_range(fluidSCproc, based_on_values = "log2Ex",
  high_quantile = 0.75, low_quantile = 0.25,
  sorted_selected_assays = NULL, groupCol = NULL, print_graphs = T)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{high_quantile}{high cutoff for gene expression range, based on quantile probabilities [0 - 1]}

\item{low_quantile}{low cutoff for gene expression range, based on quantile probabilities [0 - 1]}

\item{groupCol}{column name different sample groups, will be displayed in same plot}

\item{print_graphs}{boolean to print graphs to output}

\item{sored_selected_assays}{select subset of assays, will be displayed in same order}
}
\value{
returns a list with data and plot
}
\description{
Function to show expression range of genes based on quantile probabilities for multiple conditions
}
\details{
NA
}
\examples{
show_IQRexpression_range()
}

