% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize_fluidSC_function.R
\name{normalize_fluidSC}
\alias{normalize_fluidSC}
\title{Function to normalize fluidSCraw object}
\usage{
normalize_fluidSC(fluidSCraw, LoD = 27)
}
\arguments{
\item{fluidSCraw}{fluidSCraw S3 object}

\item{LoD}{Limit of Detection value, defaults to 26}
}
\value{
returns a fluidSCproc S3 object
}
\description{
This function processes and normalizes a fluidSCraw S3 object. It also calculates a log2 expression value based on a Limit of Detection (LoD) score
}
\details{
- if value = 999 and call = Fail, replace value with LoD (not detected)
- if value > LoD and call = Pass, replace value with LoD (not detected)
- if value < LoD and call = Fail, replace value with mean expression of assay in all cells (missing data)

Normalization is based on cell-to-cell median. Offset value between median log2 expression in sample vs average of all median values.
End result is that all cells will have the same median log2 expression.
}
\examples{
normalize_fluidSC()
}
\references{
Livak et al. "Methods for qPCR gene expression profiling applied to 1440 lymphoblastoid single cells"
}

