% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PCA_selfCluster_SC_function.R
\name{PCA_selfCluster_SC}
\alias{PCA_selfCluster_SC}
\title{Function to create a PCA analysis}
\usage{
PCA_selfCluster_SC(fluidSCproc, based_on_values = "log2Ex", PCAscaled = T,
  nr_gene_contributions = 5, colorClust = "red", firstPC = "PC1",
  secondPC = "PC2", clusterColumn = NULL, centerpoint = T, segments = T,
  corcirc = T, corGenes = T, labeling = F, corr_genes_selection = NULL,
  NAvalues = c("remove_assays", "replace_with_not_detected"),
  not_detected_value = 0, print_graphs = T)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{PCAscaled}{boolean to scale data prior to PCA or not, defaults to true}

\item{colorClust}{colors for the different cluster groups, number of colors and clusters must be the same}

\item{firstPC}{first principal component you want to visualize}

\item{secondPC}{second principal component you want to visualize}

\item{centerpoint}{boolean to show center of the different groups, defaults to TRUE}

\item{segments}{boolean to show segments on plot, defaults to TRUE}

\item{corcirc}{boolean to show correlation circle, defaults to FALSE}

\item{corGenes}{boolean to show genes that corrleate with sample position, defaults to TRUE}

\item{labeling}{boolean to show labels of samples, defaults to FALSE}

\item{corr_genes_selection}{selection of genes to see on the plot, defaults to NULL = all genes if corGenes = T}

\item{NAvalues}{choose to remove or replace assays with NA values}

\item{print_graphs}{boolean to print graphs as output}

\item{nr_gene_contribution}{number of top contributing genes in both directions to show on the PCA plot}
}
\value{
returns a list with PCAdata, PCAscores, pcaplot and screeplot
}
\description{
This function performs a PCA analysis and returns a list with the PCAresults, PCAscores, PCA plot and scree plot
}
\details{
NA
}
\examples{
PCA_selfCluster_SC()
}

