% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixplot_SC_functions.R
\name{plotCorValue_SC}
\alias{plotCorValue_SC}
\title{plotCorValue_SC()}
\usage{
plotCorValue_SC(xVar, yVar, cexText = 7.5, cexCI = 1.8,
  cor_use = "complete.obs", cor_method = "spearman")
}
\arguments{
\item{xVar,}{yVar variables for correlation score}
}
\value{
returns a spearman correlation score
}
\description{
Function calculates pairwise correlation between 2 variables
}
\details{
NA
}
\examples{
plotCorValue_SC()
}
\references{
https://hlplab.wordpress.com/2012/03/20/correlation-plot-matrices-using-the-ellipse-library/
}

