% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixplot_SC_functions.R
\name{plotMarginal_SC}
\alias{plotMarginal_SC}
\title{plotMarginal_SC function}
\usage{
plotMarginal_SC(variable, lwd = 2)
}
\arguments{
\item{variable}{vector of assay expression values}

\item{lwd}{linewidth}
}
\value{
returns a histogram with line density
}
\description{
Function creates a histogram and density line for assay for matrixplot_SC()
}
\details{
NA
}
\examples{
plotMarginal_SC()
}
\references{
https://hlplab.wordpress.com/2012/03/20/correlation-plot-matrices-using-the-ellipse-library/
}

