% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixplot_SC_functions.R
\name{plotScatter_SC}
\alias{plotScatter_SC}
\title{plotScatter_SC function}
\usage{
plotScatter_SC(dfr, xVar, yVar, cexPoints = 1.3, cexXAxis = 1.3,
  cexYAxis = 1.3, lwd = 2)
}
\arguments{
\item{dfr}{dataframe with expression values for pairwise comparison}
}
\value{
returns a scatterplot between 2 variables
}
\description{
Function creates a scatterplot for matrixplot_SC()
}
\details{
NA
}
\examples{
plotScatter_SC()
}
\references{
https://hlplab.wordpress.com/2012/03/20/correlation-plot-matrices-using-the-ellipse-library/
}

