% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_generic_methods.R
\name{print.fluidSCproc}
\alias{print.fluidSCproc}
\title{Genereric print method for S3 object fluidSCproc}
\usage{
\method{print}{fluidSCproc}(x)
}
\arguments{
\item{x}{fluidSCproc object}
}
\value{
prints first 6 lines of dataframe and additional information
}
\description{
This function prints an S3 fluidSCproc object
}
\examples{
print.fluidSCproc(x)
}
\keyword{fluidSproc}
\keyword{print}

