% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_generic_methods.R
\name{print.fluidSCraw}
\alias{print.fluidSCraw}
\title{Genereric print method for S3 object fluidSCraw}
\usage{
\method{print}{fluidSCraw}(x)
}
\arguments{
\item{x}{fluidSCraw object}
}
\value{
prints first 6 lines of dataframe and additional information
}
\description{
This function prints an S3 fluidSCraw object
}
\examples{
print.fluidSCraw(x)
}
\keyword{fluidSCraw}
\keyword{print}

