% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_Fluidigm_function.R
\name{read_Fluidigm}
\alias{read_Fluidigm}
\title{Function to create fluidSCraw object from fluidigm .csv output}
\usage{
read_Fluidigm(path_to_csv_table_results, skiplines = 11)
}
\arguments{
\item{path_to_csv_table_results}{path to the stored fluidigm results table in .csv format (e.g. "/path/to/data.csv")}

\item{skiplines}{number of lines to skip, defaults to 11}
}
\value{
Object of type fluidSCraw
}
\description{
This function takes as input the output table format from Fluidigm Data analysis and returns an S3 object 'fluidSCraw'
}
\examples{
read_Fluidigm()
}
\keyword{fluidSCraw}
\keyword{read}

