% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_function_and_methods_remove_data.R
\name{remove_data.fluidSCproc}
\alias{remove_data.fluidSCproc}
\title{Method for new generic function remove_data()}
\usage{
\method{remove_data}{fluidSCproc}(x, samples = NULL, assays = NULL)
}
\arguments{
\item{x}{S3 Object fluidSCproc}

\item{samples}{samples to remove}

\item{assays}{to remove}
}
\value{
Object of type fluidSCproc
}
\description{
This function removes Samples or Assays from fluidSCproc object
}
\examples{
remove_data(x, samples, assays)
}
\keyword{fluidSCproc}
\keyword{remove}

