% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_function_and_methods_remove_data.R
\name{remove_data.fluidSCraw}
\alias{remove_data.fluidSCraw}
\title{Method for new generic function remove_data()}
\usage{
\method{remove_data}{fluidSCraw}(x, samples = NULL, assays = NULL)
}
\arguments{
\item{x}{S3 Object fluidSCraw}

\item{samples}{samples to remove}

\item{assays}{to remove}
}
\value{
Object of type fluidSCraw
}
\description{
This function removes Samples or Assays from fluidSCraw object
}
\examples{
remove_data(x, samples, assays)
}
\keyword{fluidSCraw}
\keyword{remove}

