% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_function_and_methods_remove_data.R
\name{remove_data}
\alias{remove_data}
\title{Generic function to remove data}
\usage{
remove_data(x, samples, assays)
}
\arguments{
\item{x}{S3 Object fluidSCraw or fluidSCproc}

\item{samples}{samples to remove}

\item{assays}{to remove}
}
\value{
Object of type fluidSCraw or fluidSCproc
}
\description{
This function removes Samples or Assays from fluidSCraw or fluidSCproc object
}
\examples{
remove_data(x, samples, assays)
}
\keyword{remove}

