% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show_IQRexpression_range_function.R
\name{show_IQRexpression_range}
\alias{show_IQRexpression_range}
\title{Function to show expression range of genes based on quantile probabilities}
\usage{
show_IQRexpression_range(fluidSCproc, based_on_values = "log2Ex",
  high_quantile = 0.75, low_quantile = 0.25,
  sorted_selected_assays = NULL, groupCol = NULL)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{high_quantile}{high cutoff for gene expression range, based on quantile probabilities [0 - 1]}

\item{low_quantile}{low cutoff for gene expression range, based on quantile probabilities [0 - 1]}

\item{groupCol}{column name different sample groups, will be displayed separately}

\item{sored_selected_assays}{select subset of assays, will be displayed in same order}
}
\value{
returns a list with data and plot
}
\description{
Function to show expression range of genes based on quantile probabilities
}
\details{
NA
}
\examples{
show_IQRexpression_range()
}

