% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ternaryplot_SC_function.R
\name{ternaryplot_SC}
\alias{ternaryplot_SC}
\title{ternaryplot_SC function}
\usage{
ternaryplot_SC(fluidSCproc, based_on_values = "log2Ex",
  sorted_selected_genes = c("Pou5f1", "Psma3", "Trp53"),
  not_detected_value = 0, remove_not_detected = F)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{sorted_selected_genes}{subset of genes}

\item{not_detected_value}{value that corresponds to not detected expression}

\item{remove_not_detected}{boolean to remove not detected data}
}
\value{
returns a ternary (triangle) plot
}
\description{
Function creates a ternary or triangle plot to visualize relations between 3 variables
}
\details{
NA
}
\examples{
ternaryplot_SC()
}

