% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kmeans_cluster_estimation_functions.R
\name{variance_explained_sparseKmeans}
\alias{variance_explained_sparseKmeans}
\title{Helper function for Kmeans_cluster_explained_variance()}
\usage{
variance_explained_sparseKmeans(scores, nrClust, wbound)
}
\arguments{
\item{scores}{matrix with scores}

\item{nrClust}{number of clusters to test}

\item{wbound}{weight for kmeans clustering bound to features (genes)}
}
\value{
returns percentage of variance explained
}
\description{
Function to calculate the variance that is explained by performing sparse kmeans clustering
}

