% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/violinplotSC_function.R
\name{violinplotSC}
\alias{violinplotSC}
\title{violinplotSC()}
\usage{
violinplotSC(fluidSCproc, based_on_values = "log2Ex", assayCol = "Assays",
  assays = c("Rpl13a", "Pou5f1"), groupColumn = NULL)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{assayCol}{column name with the assays, defaults to "Assays"}

\item{assays}{vector with assays to show}

\item{groupColumn}{column name with group information, will be visualized in different facets}
}
\value{
returns a violinplot
}
\description{
Function to visualize distribution profile of assays in single cells with violin plots
}
\details{
NA
}
\examples{
violinplotSC()
}

