% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualize_distribution_function.R
\name{visualize_distribution}
\alias{visualize_distribution}
\title{Function to visualize distribution of raw or normalized data}
\usage{
visualize_distribution(fluidSCraw, based_on_values = "Value",
  mode = c("boxplot", "histogram", "density"), groupColumn = NULL)
}
\arguments{
\item{fluidSCraw}{fluidSCraw S3 object}

\item{based_on_values}{values to use for distribution visualization}

\item{mode}{visualization method, defaults to boxplot}

\item{groupColumn}{create facets based on column with grouping information}
}
\value{
creates a visualization of the distribution
}
\description{
This function visualized distribution of raw or normalized data with the use of a boxplot, histogram or density plot
}
\examples{
visualize_distribution()
}

