% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wcmds_SC_function.R
\name{wcmds_SC}
\alias{wcmds_SC}
\title{Function to create Multidimensional scaling plot}
\usage{
wcmds_SC(fluidSCproc, dimensions = 2, based_on_values = "log2Ex",
  mds_on = c("Samples", "Assays"), genes_selection = NULL,
  cluster_column_name = NULL, show_labels = T)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{mds_on}{do dimentional scaling on Samples or Assays}

\item{genes_selection}{make a selection for the genes to consider}

\item{cluster_column_name}{colors single genes based on this column}

\item{show_labels}{boolean to show gene / sample labels, defaults to TRUE}
}
\value{
returns a mds plot
}
\description{
This function creates a MDS plot
}
\details{
NA
}
\examples{
wcmds_SC()
}

