% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cluster_SC_function.R
\name{cluster_SC}
\alias{cluster_SC}
\title{Function to cluster single cell data}
\usage{
cluster_SC(fluidSCproc, based_on_values = "log2Ex", nrClust = 2,
  clustMethod = c("Kmeans", "Hierarchical", "Correlation"),
  distMethod = "euclidean", corrMethod = "pearson",
  hclustMethod = "average", selected_assays = NULL,
  cluster_column_name = "clust", nstart = 25, iter.max = 1000,
  algorithm = "Hartigan-Wong", NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{nrClust}{number of clusters you want to detect with Kmeans}

\item{clustMethod}{choose method for clustering samples}

\item{distMethod}{choose method to create distance matrix}

\item{corrMethod}{choose correlation score algorithm to use if clustMethod = Correlation}

\item{hclustMethod}{choose method to cluster distance matrix}

\item{selected_assays}{make selection of assays if wanted}

\item{cluster_column_name}{name of column that will be created with the clustering results}

\item{NAvalues}{how to handle NA values, remove or replace with not_detected_value}

\item{not_detected_value}{value to replace NA values}
}
\value{
returns a fluidSCproc S3 object appended with the cluster results
}
\description{
This function merges several clustering methods and appends the result to the fluidSCproc object
}
\details{
NA
}
\examples{
cluster_SC()
}

