% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ellipseCorrPlot_SC_function.R
\name{ellipseCorrPlot_SC}
\alias{ellipseCorrPlot_SC}
\title{Function to visualize correlation between assay expression}
\usage{
ellipseCorrPlot_SC(fluidSCproc, based_on_values = "log2Ex",
  sorted_selected_genes = c("Pou5f1", "Psma3", "Trp53", "Nanog"),
  mycol = colorRamp(c("#CC0000", "white", "#3366CC"), space = "Lab"),
  outline = FALSE, col = "grey", upper.panel = c("ellipse", "number",
  "none"), lower.panel = c("ellipse", "number", "none"), diag = c("none",
  "ellipse", "number"), digits = 2, bty = "n", axes = FALSE, xlab = "",
  ylab = "", asp = 1, cex.lab = par("cex.lab"), cex = 0.75 * par("cex"),
  mar = 0.1 + c(2, 2, 4, 2), ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{sorted_selected_genes}{select genes for pair-wise correlation, will be displayed in same order}

\item{mycol}{color to use}

\item{outline}{outline ellipse or not, defaults to FALSE}

\item{upper.panel}{information to put in upper panels}

\item{lower.panel}{information to put in lower panels}

\item{diag}{information to put in diagonal}

\item{...}{}
}
\value{
returns a ellipse scatterplot with correlation that are translated in ellipse images
}
\description{
Function creates a visualization for assay correlation scores based on ellipse package
}
\details{
NA
}
\examples{
ellipseCorrPlot_SC()
}
\references{
https://hlplab.wordpress.com/2012/03/20/correlation-plot-matrices-using-the-ellipse-library/
}

