% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmapSC_function.R
\name{heatmapSC}
\alias{heatmapSC}
\title{Function to create a heatmap from single cell data}
\usage{
heatmapSC(fluidSCproc, based_on_values = "log2Ex",
  distance_method = c("pearson", "spearman", "euclidean", "maximum",
  "manhatten", "binary"), clust_method = c("average", "complete", "ward.D",
  "ward.D2", "single", "median"), NAvalues = c("remove_assays",
  "replace_with_not_detected"), not_detected_value = 0, hmcolor = NULL, ...)
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object~}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{distance_method}{choose method to calculute distance matrix: pearson, spearman, euclidean, maxium, manhatten, binary}

\item{clust_method}{choose method to cluster distance matrix: average, complete, ward.D, ward.D2, single, median}

\item{NAvalues}{how to process NA values, remove or replace with not_detected_value (default = 0)}

\item{not_detected_value}{value to replace NA values with}

\item{hmcolor}{give your own colors or colorpalette}
}
\value{
returns a heatmap plot
}
\description{
This function creates a heatmap?
}
\details{
NA
}
\examples{
heatmapSC()
}

