% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matrixplot_SC_functions.R
\name{matrixplot_SC}
\alias{matrixplot_SC}
\title{matrixplot_SC()}
\usage{
matrixplot_SC(fluidSCproc, based_on_values = "log2Ex",
  sorted_selected_genes = c("Pou5f1", "Psma3", "Trp53", "Nanog", "Cdx2"))
}
\arguments{
\item{fluidSCproc}{fluidSCproc S3 object}

\item{based_on_values}{values to use, defaults to "log2Ex"}

\item{sorted_selected_genes}{subset of genes to select for visualization}
}
\value{
returns a histogram with line density
}
\description{
Function creates a pairwise matrixplot between assays and shows correlation, distribution and scatterplot
}
\details{
NA
}
\examples{
matrixplot_SC()
}
\references{
https://hlplab.wordpress.com/2012/03/20/correlation-plot-matrices-using-the-ellipse-library/
}

